third_party/blink/web_tests/svg/dom/text-rotate-live-expected.txt

This tests that the SVGAnimatedNumberList returned by SVGTextPositioningElement.rotate is live.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getRotate() is '10,20,30'
PASS getAndSetRotate('40 50') is '40,50'
PASS getAndSetRotate('6e1') is '60'
PASS getAndSetRotate('-.5  , 40E-0') is '-0.5,40'
PASS successfullyParsed is true

TEST COMPLETE
